+++ src/main/java/com/wheelinspection/service/BreakdownHistoryService.java
@@ -2,13 +2,16 @@ package com.wheelinspection.service;
 
 
 import com.wheelinspection.entity.BreakdownHistory;
+import com.wheelinspection.handler.ServiceException;
 import com.wheelinspection.repository.BreakdownHistoryRepository;
+import com.wheelinspection.responce.ApplicationResponce;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class BreakdownHistoryService {
@@ -20,17 +23,39 @@ public class BreakdownHistoryService {
         return repository.findAll();
     }
 
-    public BreakdownHistory getRecordById(Long id) {
-        return repository.findById(id).orElse(null);
+    public ApplicationResponce getRecordById(Long id) throws ServiceException {
+       Optional<BreakdownHistory> breakdownHistory =repository.findById(id);
+       if(breakdownHistory.isPresent()){
+           ApplicationResponce applicationResponce =new ApplicationResponce();
+           applicationResponce.setData(breakdownHistory);
+           return applicationResponce;
+       }else{
+           throw new ServiceException("record not found",34);
+       }
     }
 
-    public BreakdownHistory addOrUpdateRecord(BreakdownHistory record) {
-        return repository.save(record);
+    public ApplicationResponce addOrUpdateRecord(BreakdownHistory record) throws ServiceException {
+
+        BreakdownHistory breakdownHistory = repository.save(record);
+        if(breakdownHistory!=null){
+            ApplicationResponce applicationResponce =new ApplicationResponce();
+            applicationResponce.setData("Save or Update successfully");
+            return applicationResponce;
+        }else{
+            throw new ServiceException("not save or update successfully",103);
+        }
     }
 
-    public String deleteRecord(Long id) {
-        repository.deleteById(id);
-        return "delete successfully";
+    public ApplicationResponce deleteRecord(Long id) throws ServiceException {
+        if(repository.existsById(id)){
+            repository.deleteById(id);
+            ApplicationResponce applicationResponce  =new ApplicationResponce();
+            applicationResponce.setData("delete successfully");
+            return applicationResponce;
+        }else{
+            throw new ServiceException("Record not found",102);
+        }
+
     }
 
     public Page<BreakdownHistory> getPaginatedData(String search, Pageable pageable) {
