+++ src/main/java/com/wheelinspection/service/DemuBearingRejectionDataService.java
@@ -2,7 +2,9 @@ package com.wheelinspection.service;
 
 import com.wheelinspection.entity.CondemenedDisposal;
 import com.wheelinspection.entity.DemuBearingRejectionData;
+import com.wheelinspection.handler.ServiceException;
 import com.wheelinspection.repository.DemuBearingRejectionDataRepository;
+import com.wheelinspection.responce.ApplicationResponce;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -19,25 +21,51 @@ public class DemuBearingRejectionDataService {
         return repository.findAll();
     }
 
-    public Optional<DemuBearingRejectionData> getRecordById(Long id) {
-        return repository.findById(id);
+    public ApplicationResponce getRecordById(Long id) throws ServiceException {
+        if (repository.existsById(id)) {
+            Optional<DemuBearingRejectionData> demuBearingRejectionData = repository.findById(id);
+            if (demuBearingRejectionData.isPresent()) {
+                ApplicationResponce applicationResponce = new ApplicationResponce();
+                applicationResponce.setData(demuBearingRejectionData);
+                return applicationResponce;
+            }
+
+    }
+        throw new ServiceException("record not found",23);
     }
 
-    public DemuBearingRejectionData createRecord(DemuBearingRejectionData record) {
-        return repository.save(record);
+    public ApplicationResponce createRecord(DemuBearingRejectionData record) throws ServiceException {
+        DemuBearingRejectionData demuBearingRejectionData=  repository.save(record);
+        if(demuBearingRejectionData!=null){
+            ApplicationResponce applicationResponce  =new ApplicationResponce();
+            applicationResponce.setData(demuBearingRejectionData);
+            return applicationResponce;
+        }else{
+            throw new ServiceException("Record not save successfully ",102);
+        }
     }
 
-    public DemuBearingRejectionData updateRecord(Long id, DemuBearingRejectionData updatedRecord) {
+    public ApplicationResponce updateRecord(Long id, DemuBearingRejectionData updatedRecord) throws ServiceException {
         if (repository.existsById(id)) {
             updatedRecord.setId(id);
-            return repository.save(updatedRecord);
+            DemuBearingRejectionData demuBearingRejectionData = repository.save(updatedRecord);
+            if(demuBearingRejectionData!=null){
+                ApplicationResponce applicationResponce =new ApplicationResponce();
+                applicationResponce.setData("update record successfully");
+                return applicationResponce;
+            }
         }
-        return null;
+        throw new ServiceException("record not found ",98);
     }
 
-    public String deleteRecord(Long id) {
-        repository.deleteById(id);
-        return "delete successfully";
+    public ApplicationResponce deleteRecord(Long id) throws ServiceException {
+        if(repository.existsById(id)){
+            repository.deleteById(id);
+            ApplicationResponce applicationResponce=new ApplicationResponce();
+            applicationResponce.setData("delete successfully");
+            return applicationResponce;
+        }
+       throw new ServiceException("record not found",98);
     }
