+++ src/main/java/com/wheelinspection/service/PreInspectionRecordService.java
@@ -2,7 +2,9 @@ package com.wheelinspection.service;
 
 import com.wheelinspection.entity.Machine;
 import com.wheelinspection.entity.PreInspectionRecord;
+import com.wheelinspection.handler.ServiceException;
 import com.wheelinspection.repository.PreInspectionRecordRepository;
+import com.wheelinspection.responce.ApplicationResponce;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -19,25 +21,52 @@ public class PreInspectionRecordService {
         return repository.findAll();
     }
 
-    public Optional<PreInspectionRecord> getRecordById(Long id) {
-        return repository.findById(id);
+    public ApplicationResponce getRecordById(Long id) throws ServiceException {
+        if(repository.existsById(id)){
+           Optional<PreInspectionRecord>  preInspectionRecord  = repository.findById(id);
+           if(preInspectionRecord.isPresent()){
+               ApplicationResponce applicationResponce =new ApplicationResponce();
+               applicationResponce.setData(preInspectionRecord);
+               return applicationResponce;
+           }
+        }
+            throw new ServiceException("Record not found",404);
     }
 
-    public PreInspectionRecord createRecord(PreInspectionRecord record) {
-        return repository.save(record);
+    public ApplicationResponce createRecord(PreInspectionRecord record) throws ServiceException {
+        PreInspectionRecord preInspectionRecord  =repository.save(record);
+        if(preInspectionRecord!=null){
+            ApplicationResponce applicationResponce  =new ApplicationResponce();
+            applicationResponce.setData("save record successfully");
+            return applicationResponce;
+        }else{
+            throw new ServiceException("not save successfully",500);
+        }
     }
 
-    public PreInspectionRecord updateRecord(Long id, PreInspectionRecord updatedRecord) {
+    public ApplicationResponce updateRecord(Long id, PreInspectionRecord updatedRecord) throws ServiceException {
         if (repository.existsById(id)) {
             updatedRecord.setId(id);
-            return repository.save(updatedRecord);
+            PreInspectionRecord preInspectionRecord = repository.save(updatedRecord);
+            if(preInspectionRecord!=null){
+                ApplicationResponce applicationResponce =new ApplicationResponce();
+                applicationResponce.setData("Update successfully");
+                return applicationResponce;
+            }
         }
-        return null;
+        throw new ServiceException("Record Not found",500);
     }
 
-    public String deleteRecord(Long id) {
-        repository.deleteById(id);
-        return "Delete successfully";
+    public ApplicationResponce deleteRecord(Long id) throws ServiceException {
+        if(repository.existsById(id)){
+            repository.deleteById(id);
+            ApplicationResponce applicationResponce=new ApplicationResponce();
+            applicationResponce.setData("Delete successfully");
+            return applicationResponce;
+        }
+        else{
+            throw new ServiceException("Record not found",404);
+        }
     }
