+++ src/main/java/com/wheelinspection/service/Sheet22Service.java
@@ -2,7 +2,9 @@ package com.wheelinspection.service;
 
 import com.wheelinspection.entity.Rard;
 import com.wheelinspection.entity.Sheet22;
+import com.wheelinspection.handler.ServiceException;
 import com.wheelinspection.repository.Sheet22Repository;
+import com.wheelinspection.responce.ApplicationResponce;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -19,25 +21,49 @@ public class Sheet22Service {
         return repository.findAll();
     }
 
-    public Optional<Sheet22> getRecordById(Long id) {
-        return repository.findById(id);
+    public ApplicationResponce getRecordById(Long id) throws ServiceException {
+        if(repository.existsById(id)) {
+            Optional<Sheet22> sheet22 = repository.findById(id);
+            if(sheet22.isPresent()){
+                ApplicationResponce applicationResponce   =new ApplicationResponce();
+                applicationResponce.setData(sheet22);
+                return applicationResponce;
+            }
+        }
+        throw new ServiceException("Record not found",404);
     }
 
-    public Sheet22 createRecord(Sheet22 record) {
-        return repository.save(record);
+    public ApplicationResponce createRecord(Sheet22 record) throws ServiceException {
+        Sheet22 sheet22 = repository.save(record);
+        if(sheet22!=null){
+            ApplicationResponce applicationResponce=new ApplicationResponce();
+            applicationResponce.setData("Record save Successfully");
+            return applicationResponce;
+        }
+        throw new ServiceException("not save sucessfully",500);
     }
 
-    public Sheet22 updateRecord(Long id, Sheet22 updatedRecord) {
+    public ApplicationResponce updateRecord(Long id, Sheet22 updatedRecord) throws ServiceException {
         if (repository.existsById(id)) {
             updatedRecord.setId(id);
-            return repository.save(updatedRecord);
+            Sheet22 sheet22 = repository.save(updatedRecord);
+            if(sheet22!=null){
+                ApplicationResponce applicationResponce=new ApplicationResponce();
+                applicationResponce.setData("Record save Successfully");
+                return applicationResponce;
+            }
         }
-        return null;
+        throw new ServiceException("Record not found",404);
     }
 
-    public String deleteRecord(Long id) {
-        repository.deleteById(id);
-        return "delete successfully";
+    public ApplicationResponce deleteRecord(Long id) throws ServiceException {
+        if(repository.existsById(id)){
+            repository.deleteById(id);
+            ApplicationResponce applicationResponce=new ApplicationResponce();
+            applicationResponce.setData("delete successfully");
+            return applicationResponce;
+        }
+         throw new ServiceException("",404);
     }
